Moves pages folder out of components folder so it's directly under src, as per common convention
